Added some more classes
Have to hook up more things,
Some additions to the button class.
I have to work buttons into the button manager,
I still have to work in the text input.
There is a lot to be done I’ll be working on it starting Tuesday
through due date, this is not really the commit i wanted to make but
we’re punched for time.